Rename types module to custom_types
