+++ tests/util/test_util_hooks.py
@@ -3,8 +3,8 @@ from unittest import mock
 import pytest
 
 import tfworker.util.hooks as hooks
+from tfworker.custom_types.terraform import TerraformAction, TerraformStage
 from tfworker.exceptions import HookError
-from tfworker.types.terraform import TerraformAction, TerraformStage
 
 
 # Fixture for a mock Terraform state file
