+++ tfworker/backends/s3.py
@@ -11,8 +11,8 @@ import click
 
 import tfworker.util.log as log
 import tfworker.util.terraform as tf_util
+from tfworker.custom_types import JSONType
 from tfworker.exceptions import BackendError
-from tfworker.types import JSONType
 
 from .base import BaseBackend, validate_backend_empty
