+++ tfworker/custom_types/freezable_basemodel.py
