+++ tfworker/handlers/sqs.py
@@ -32,8 +32,8 @@ import click
 from pydantic import BaseModel, Field
 
 import tfworker.util.log as log
+from tfworker.custom_types.terraform import TerraformAction, TerraformStage
 from tfworker.exceptions import HandlerError
-from tfworker.types.terraform import TerraformAction, TerraformStage
 
 from .base import BaseHandler
 from .registry import HandlerRegistry
